john_project_test/pyproject.toml
[tool.poetry]
name = "hao-project-test"
version = "0.1.0"
description = ""
authors = ["SMH642800 <SMH642800@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
google-cloud-vision = "^3.4.4"
google-cloud-translate = "^3.12.0"
pyside6 = "^6.5.2"
opencv-python = "^4.8.1.78"
pyautogui = "^0.9.52"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
